fix: delete team

+++ server/controller/teams.js
@@ -1,4 +1,5 @@
 const Teams = require("../model/teams.js");
+const Category = require("../model/category.js");
 const toDelete = require("../middleware/toDelete.js");
 
 module.exports.create = async (req, res) => {
@@ -24,19 +25,24 @@ module.exports.find = async (req, res) => {
 
 module.exports.update = async (req, res) => {
   try {
-    const id = req.params.id;
-    const { name, img } = req.body;
-    if (!name || !img)
-      return res.status(400).json({ msg: "Body is not valid" });
+    if (req.file) {
+      const id = req.params.id;
+      const { img } = await Teams.findById(id);
+      const { name } = req.body;
+      if (!name) return res.status(400).json({ msg: "Body is not valid" });
 
-    await Teams.findByIdAndUpdate(id, {
-      $set: {
-        name: name,
-        img: img,
-      },
-    });
+      await Teams.findByIdAndUpdate(id, {
+        $set: {
+          name: name,
+          img: req.file.filename,
+        },
+      });
+      
+      toDelete(img);
 
-    res.status(200).json({ msg: "Team has been changed" });
+      res.status(200).json({ msg: "Team has been changed" });
+    }
+    res.status(400).json({ msg: "Team has not been changed" });
   } catch (error) {
     res.status(400).json({ msg: "Can not be changed" });
   }
@@ -45,6 +51,13 @@ module.exports.update = async (req, res) => {
 module.exports.delete = async (req, res) => {
   try {
     const id = req.params.id;
+
+    const team = await Category.findOne({
+      $or: [{ "teams.team1": id }, { "teams.team2": id }],
+    });
+
+    if (team) return res.json({ msg: "Team can not be deleted" });
+
     const { img } = await Teams.findById(id);
     toDelete(img);
     await Teams.findByIdAndDelete(id);
